170320 18:38 // JJH #Board RENEW

+++ sugo2day/src/main/java/member/login/AuthInfo.java
@@ -3,19 +3,19 @@ package member.login;
 // 로그인 성공 후, 인증 상태 정보를 세션에 보관
 public class AuthInfo {
 	
-	private Long id;
+	private int id;
 	private String email;
 	private String name;
 	
 
-	public AuthInfo(Long id, String email, String name) {
+	public AuthInfo(int id, String email, String name) {
 		this.id = id;
 		this.email = email;
 		this.name = name;
 	}
 	
 	
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
